Module iota_rpc_loadgen::payload::rpc_command_processor

Structs

RpcCommandProcessor

Enums

CacheType

Constants

DEFAULT_GAS_BUDGET 🔒

DEFAULT_LARGE_GAS_BUDGET 🔒

MAX_NUM_NEW_OBJECTS_IN_SINGLE_TRANSACTION 🔒

Functions

calculate_split_amounts 🔒

Calculate the split amounts for a given number of coins, amount per coin, and maximum number of coins per transaction. Returns a Vec of (primary_coin_amount, split_into_n_coins)

divide_checkpoint_tasks 🔒

divide_get_all_balances_tasks 🔒

divide_get_object_tasks 🔒

divide_multi_get_objects_tasks 🔒

divide_multi_get_transaction_blocks_tasks 🔒

divide_query_transaction_blocks_tasks 🔒

get_coin_with_balance 🔒

get_coin_with_max_balance 🔒

get_iota_coin_ids 🔒

load_addresses_from_file

load_digests_from_file

load_objects_from_file

num_transactions_needed 🔒

Calculate the number of transactions needed to split the given number of coins. new_coins_per_txn must be greater than 0

pay_iota 🔒

prepare_new_signer_and_coins 🔒

read_data_from_file 🔒

sign_and_execute 🔒

split_coins 🔒

write_data_to_file 🔒